Move interfaces to a separate file
